E2EEvaluation/requirements.txt
matplotlib
numpy
FreeSimpleGUI
PyYAML
scipy
tikzplotlib-patched @ git+https://github.com/JasonGross/tikzplotlib.git@74b820f241f688889195e55442000672d1526fd0